Create rental, book,list_rentals in separate files
